requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
email-validator
requests
